requirements.txt
wheel==0.40.0
pytest==7.4.0
PyPDF2==3.0.1
pdfminer.six==20221105
python-Levenshtein==0.21.1
scikit-learn==1.3.0
pytesseract==0.3.10
Pillow>=10.0.0
pymupdf>=1.17.0
pdf2image==1.16.3
pandas==2.0.3
torch>=2.0.1 --index-url https://download.pytorch.org/whl/cu118
transformers>=4.30.2
huggingface-hub>=0.16.4
matplotlib>=3.7.2
timm>=0.9.2
python-docx==0.8.11
chardet==5.2.0
extract-msg
python-pptx
striprtf
docx2pdf
fpdf
comtypes

#Used To read excel files with pandas
openpyxl>=3.1.2

#Used to generate fake information for the GeneratedTable.py
faker==2.0.5

#Generate a random dataframe
pydbgen==1.0.5

#Used to write Latex using python
pylatex==1.3.4

#Used to create filenames for GeneratedTable.py
shortuuid==1.0.11

#Used to write Excel files
xlsxwriter==3.1.2

# Used for the GUI
dash
dash[diskcache]
celery

# for exe
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
